requirements.txt
# ── Web Framework ──────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Database ────────────────────────────────────
mysql-connector-python==9.1.0

# ── Vector Store (choose ONE - see instructions) 
# Option A: CPU (always works)
faiss-cpu==1.9.0
# Option B: GPU (only if CUDA is confirmed)
# faiss-gpu==1.7.2

# ── Embeddings ──────────────────────────────────
sentence-transformers==3.2.1
numpy==1.26.4

# ── PDF Ingestion ───────────────────────────────
pypdf==5.1.0

# ── URL Ingestion ───────────────────────────────
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# ── YouTube Ingestion ───────────────────────────
youtube-transcript-api==1.2.4

# ── LLM (Ollama) ────────────────────────────────
httpx==0.27.2

pdfplumber==0.11.4

groq

rank-bm25

torch

python-dotenv